Currently working on several additions...
An user interface for the player, so that health and stamina values can be indicated visually. A restart level screen, if the player dies, a screen will appear prompting the player whether they want to restart the level or quit the game. A random enemy will be selected to drop the key that the player uses complete the level.